index.js: bug fix: two peers as initiators
Fix bug occurring when two peers try to connect to each other both as
initiators.
Each peer generates a random value and only connects to peer that have
generated a lower value. So when two peers connect to each other one peer
(that has the lower value) will ignore the 'connect' message and the other
(that has the higher value) will connect as initiator.